+++ src/sound/adpcm.h
@@ -48,7 +48,7 @@ int OKIM6295_sh_start(const struct MachineSound *msound);
 void OKIM6295_sh_stop(void);
 void OKIM6295_sh_update(void);
 void OKIM6295_set_bank_base(int which, int base);
-void OKIM6295_set_frequency(int which, int frequency);
+void OKIM6295_set_frequency(int which, double frequency);
 
 READ_HANDLER( OKIM6295_status_0_r );
 READ_HANDLER( OKIM6295_status_1_r );
